[Fix] fileをencodeするロジックを移動
